requirements.txt
asgiref==3.7.2
blinker==1.6.2
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
Django==4.2.3
djangorestframework==3.14.0
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
sseclient==0.0.27
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
